Add default unit testing support to the Figwheel template.
The template now includes the necessary configuration in project.clj,
plus supporting files, to allow the user to execute ```lein cljsbuild
once test``` and have unit tests run in PhantomJS< if it's installed.

Running the unit tests naturally loads the core.cljs, which prior to my
modifications caused Reagent / Om (if used) to trigger, because they were
being called directly from the core NS.

I've changed core.cljs and index.html to
explicitly call a main function when we're running normally - and not to
call it when we're testing (unless the user expliclty does so) which
prevents Om / Reagent erroring because they can't find the app root
when we're running tests.

(The alternative would have been to add a <div id="app"> to the
unit-test.html, but at that point we're taking control of where to root
the app away from the developer)